laravel to cache files
Not sure if you’ve noticed, but opening an app or visiting a website for the first time can take a while.

Okay, a “while” may be extreme, but those extra few seconds of loading time matter to us when surfing the web or scrolling through our apps.

Although, it isn’t a coincidence why this happens. It’s mainly due to our devices not having cached data for a particular app or website.

If you’re unsure what cached data is, let's take a quick look.